requirements.txt
# Flask Application Requirements
# Shopify Product Automation

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# HTTP Requests
requests==2.31.0

# OpenAI API
openai==1.3.0

# Google APIs
# google-genai is the NEW SDK for Nano Banana (Gemini 2.5 Flash Image)
google-genai>=0.3.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Image Processing
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2

# Production Server (optional)
gunicorn==21.2.0
